@@ -16,16 +16,19 @@ repositories {
 }
 
 ext {
-    guiceVersion = "4.1.0"
+    guiceVersion = "4.2.2"
     log4jVersion = "2.8.+"
     jacksonVersion = "2.9.+"
+    ver_junit = "5.3.1"
+
     dep_owner = "org.aeonbits.owner:owner:1.0.10"
     dep_rxjava = "io.reactivex.rxjava2:rxjava:2.1.+"
     dep_jgroups = "org.jgroups:jgroups:4.0.13.Final"
     dep_inject = "javax.inject:javax.inject:1"
     dep_rabbitmq = "com.rabbitmq:amqp-client:5.5.0"
     dep_guice = "com.google.inject.extensions:guice-multibindings:${guiceVersion}"
-    dep_testcontainers = "org.testcontainers:testcontainers:1.9.1"
+    dep_testcontainers_junit = "org.testcontainers:junit-jupiter:1.10.1"
+    dep_testcontainers = "org.testcontainers:testcontainers:1.10.1"
 
     proj_test_util = "clustercode.test.util"
     proj_api_transcode = "clustercode.api.transcode"
@@ -61,8 +64,8 @@ subprojects {
     group = 'clustercode'
     version = '1.0'
 
-    targetCompatibility = 1.8
-    sourceCompatibility = 1.8
+    targetCompatibility = 1.10
+    sourceCompatibility = 1.10
 
     apply plugin: "java"
 
@@ -88,7 +91,8 @@ subprojects {
 
         // testing
         testRuntime "org.apache.logging.log4j:log4j-slf4j-impl:${log4jVersion}"
-        testCompile "junit:junit:4.12"
+        testRuntime "org.junit.jupiter:junit-jupiter-engine:${ver_junit}"
+        testCompile "org.junit.jupiter:junit-jupiter-api:${ver_junit}"
         testCompile "org.mockito:mockito-all:2.0.+"
         testCompile "com.google.jimfs:jimfs:1.+"
         testCompile "org.assertj:assertj-core:3.8.0"
@@ -102,7 +106,7 @@ subprojects {
 
 }
 
-dependencies{
+dependencies {
     compile project(":${proj_main}")
 }
 
@@ -122,7 +126,7 @@ shadowJar {
 task fullBuild(type: Jar) {
     manifest {
         attributes "Main-Class": "clustercode.main.Startup",
-            "Implementation-Title": project.name
+                "Implementation-Title": project.name
     }
     archiveName = "clustercode.jar"
     baseName = project.name
