@@ -5,8 +5,8 @@ import clustercode.api.domain.Media;
 import clustercode.api.event.messages.TranscodeFinishedEvent;
 import clustercode.impl.cleanup.CleanupConfig;
 import clustercode.test.util.FileBasedUnitTest;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.mockito.Spy;
@@ -16,7 +16,7 @@ import java.nio.file.Path;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
-public class DeleteSourceProcessorTest implements FileBasedUnitTest {
+public class DeleteSourceProcessorTest {
 
     private DeleteSourceProcessor subject;
     private Path inputDir;
@@ -30,12 +30,13 @@ public class DeleteSourceProcessorTest implements FileBasedUnitTest {
     @Spy
     private Media media;
 
-    @Before
+    private FileBasedUnitTest fs = new FileBasedUnitTest();
+
+    @BeforeEach
     public void setUp() throws Exception {
         MockitoAnnotations.initMocks(this);
-        setupFileSystem();
 
-        inputDir = getPath("input");
+        inputDir = fs.getPath("input");
         when(cleanupConfig.base_input_dir()).thenReturn(inputDir);
         transcodeFinishedEvent.setMedia(media);
         context.setTranscodeFinishedEvent(transcodeFinishedEvent);
@@ -45,7 +46,7 @@ public class DeleteSourceProcessorTest implements FileBasedUnitTest {
     @Test
     public void processStep_ShouldDeleteSourceFile_IfFileExists() throws Exception {
 
-        Path source = createFile(getPath("0", "video.ext"));
+        var source = fs.createFile(fs.getPath("0", "video.ext"));
         media.setSourcePath(source);
 
         subject.processStep(context);
@@ -55,7 +56,7 @@ public class DeleteSourceProcessorTest implements FileBasedUnitTest {
 
     @Test
     public void processStep_ShouldDoNothing_IfFileNotExists() throws Exception {
-        Path source = getPath("0", "video.ext");
+        var source = fs.getPath("0", "video.ext");
         media.setSourcePath(source);
 
         subject.processStep(context);
