@@ -5,8 +5,8 @@ import clustercode.api.domain.Media;
 import clustercode.api.event.messages.TranscodeFinishedEvent;
 import clustercode.impl.cleanup.CleanupConfig;
 import clustercode.test.util.FileBasedUnitTest;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.mockito.Spy;
@@ -16,7 +16,7 @@ import java.nio.file.Path;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
-public class MarkSourceProcessorTest implements FileBasedUnitTest {
+public class MarkSourceProcessorTest {
 
     private MarkSourceProcessor subject;
     private Path inputDir;
@@ -30,12 +30,13 @@ public class MarkSourceProcessorTest implements FileBasedUnitTest {
     @Spy
     private Media media;
 
-    @Before
+    private FileBasedUnitTest fs = new FileBasedUnitTest();
+
+    @BeforeEach
     public void setUp() throws Exception {
         MockitoAnnotations.initMocks(this);
-        setupFileSystem();
 
-        inputDir = getPath("input");
+        inputDir = fs.getPath("input");
         context.setTranscodeFinishedEvent(transcodeFinishedEvent);
         transcodeFinishedEvent.setMedia(media);
         transcodeFinishedEvent.setSuccessful(true);
@@ -47,10 +48,10 @@ public class MarkSourceProcessorTest implements FileBasedUnitTest {
 
     @Test
     public void processStep_ShouldCreateFile_IfSourceDoesExist() throws Exception {
-        Path source = createFile(getPath("0", "video.ext"));
-        Path expected = inputDir.resolve("0").resolve("video.ext.done");
+        var source = fs.createFile(fs.getPath("0", "video.ext"));
+        var expected = inputDir.resolve("0").resolve("video.ext.done");
 
-        createFile(inputDir.resolve(source));
+        fs.createFile(inputDir.resolve(source));
         media.setSourcePath(source);
         subject.processStep(context);
 
@@ -59,8 +60,8 @@ public class MarkSourceProcessorTest implements FileBasedUnitTest {
 
     @Test
     public void processStep_ShouldNotCreateFile_IfSourceDoesNotExist() throws Exception {
-        Path source = getPath("0", "video.ext");
-        Path expected = getPath("0", "video.ext.done");
+        var source = fs.getPath("0", "video.ext");
+        var expected = fs.getPath("0", "video.ext.done");
 
         media.setSourcePath(source);
         subject.processStep(context);
