@@ -2,8 +2,8 @@ package clustercode.impl.scan;
 
 import clustercode.api.domain.Profile;
 import clustercode.test.util.FileBasedUnitTest;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -13,45 +13,44 @@ import java.util.Map;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.data.Index.atIndex;
 
-public class ProfileParserImplTest implements FileBasedUnitTest {
+public class ProfileParserImplTest {
 
     private ProfileParserImpl subject;
+    private FileBasedUnitTest fs = new FileBasedUnitTest();
 
-    @Before
+    @BeforeEach
     public void setUp() throws Exception {
         subject = new ProfileParserImpl();
-        setupFileSystem();
     }
 
     @Test
     public void parseFile_ShouldIgnoreEmptyLine() throws Exception {
-        Path testFile = getPath("profile.ffmpeg");
-        String option1 = " option with space";
-        String option2 = "another line";
+        Path testFile = fs.getPath("profile.ffmpeg");
+        String option1 = " line_without_space";
+        String option2 = "line with space";
         Files.write(testFile, Arrays.asList(option1, "", option2));
 
         List<String> results = subject.parseFile(testFile).get().getArguments();
 
-        assertThat(results).containsExactly("option", "with", "space", "another", "line");
+        assertThat(results).containsExactly("line_without_space", "line with space");
     }
 
     @Test
     public void parseFile_ShouldIgnoreFieldLines() throws Exception {
-        Path testFile = getPath("profile.ffmpeg");
+        Path testFile = fs.getPath("profile.ffmpeg");
         String option1 = " %{FIELD=value}";
         String option2 = "another line";
         Files.write(testFile, Arrays.asList(option1, option2));
 
         List<String> results = subject.parseFile(testFile).get().getArguments();
 
-        assertThat(results).containsExactly("another", "line");
+        assertThat(results).containsExactly("another line");
     }
 
     @Test
     public void parseFile_ShouldParseFieldLines() throws Exception {
-        Path testFile = getPath("profile.ffmpeg");
+        Path testFile = fs.getPath("profile.ffmpeg");
         String option1 = "%{FIELD=value}";
         String option2 = "%{key=other}";
         Files.write(testFile, Arrays.asList(option1, option2));
@@ -59,14 +58,14 @@ public class ProfileParserImplTest implements FileBasedUnitTest {
         Map<String, String> results = subject.parseFile(testFile).get().getFields();
 
         assertThat(results)
-                .containsKeys("FIELD", "KEY")
-                .containsValues("value", "other")
-                .hasSize(2);
+            .containsKeys("FIELD", "KEY")
+            .containsValues("value", "other")
+            .hasSize(2);
     }
 
     @Test
     public void parseFile_ShouldReturnEmptyProfile_OnError() throws Exception {
-        Path testFile = getPath("profile.ffmpeg");
+        Path testFile = fs.getPath("profile.ffmpeg");
 
         Optional<Profile> result = subject.parseFile(testFile);
 
@@ -137,29 +136,5 @@ public class ProfileParserImplTest implements FileBasedUnitTest {
     public void extractValue_ShouldReturnEmptyString_IfNoValuePresent() throws Exception {
         String testLine = "%{KEY=}";
         assertThat(subject.extractValue(testLine)).isEqualTo("");
-    }
-
-    @Test
-    public void separateWhitespace_ShouldReturnTwoElements() throws Exception {
-        String testLine = "one two";
-        assertThat(subject.separateWhitespace(testLine))
-                .contains("one", atIndex(0))
-                .contains("two", atIndex(1))
-                .hasSize(2);
-    }
-
-    @Test
-    public void separateWhitespace_ShouldReturnOneElement_IfNoWhitespace() throws Exception {
-        String testLine = "one";
-        assertThat(subject.separateWhitespace(testLine))
-                .contains("one", atIndex(0))
-                .hasSize(1);
-    }
-    @Test
-    public void separateWhitespace_ShouldReturnOneElement_IfWhitespaceIsQuoted() throws Exception {
-        String testLine = "\"one two \"";
-        assertThat(subject.separateWhitespace(testLine))
-                .contains("one two ", atIndex(0))
-                .hasSize(1);
     }
 }
