@@ -2,11 +2,10 @@ package clustercode.test.integration;
 
 import clustercode.api.rest.v1.dto.Task;
 import lombok.extern.slf4j.Slf4j;
-import lombok.var;
-import org.junit.BeforeClass;
-import org.junit.Ignore;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
 import javax.ws.rs.core.GenericType;
 import java.io.IOException;
@@ -18,13 +17,14 @@ import java.util.concurrent.TimeoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @Slf4j
+@Testcontainers
 public class TaskListIT {
 
-    @Rule
-    public ClustercodeContainer instance1 = new ClustercodeContainer(7700)
+    @Container
+    private static ClustercodeContainer instance1 = new ClustercodeContainer(7700)
             .withOutputLogger(log);
 
-    @BeforeClass
+    @BeforeAll
     public static void setup() throws IOException {
         var target = Paths.get("input", "0", "120s_blank.mp4");
         var doneFile = target.resolveSibling("120s_blank.mp4.done");
@@ -32,7 +32,7 @@ public class TaskListIT {
     }
 
     @Test
-    @Ignore("not working atm")
+    //@Disabled("not working atm")
     public void taskList_ShouldBeEmpty_WhenNoneFound() throws TimeoutException, InterruptedException {
         log.info("Waiting...");
         instance1.waitUntilLineStartsWith("ExternalProcess - Invoking: [/usr/bin/ffmpeg,");
